working...

+++ Core/Src/main.c
@@ -54,7 +54,7 @@
 
 /* USER CODE BEGIN PV */
 uint16_t adcBuffer[ADC_BUFFER_SIZE];
-volatile uint16_t ADC_Value[ADC_BUFFER_SIZE]; // 声明数组来存储ADC采样结果
+volatile float ADC_Value[ADC_BUFFER_SIZE]; // 声明数组来存储ADC采样结果
 uint16_t filtered_adc_values[7]; // 过滤后的ADC
 uint16_t filtered_voltage[7];     // 过滤后的电压
